Fixed diagonalization error
There was a small error causing covariance matrices in the last part of the script to not be diagonalized appropriately (i.e. turning variances into a diagonal cov mtx). This led to between-observation heterogeneity values of less than 1, which can be observed in the figure showing the RRH of latent space samples. Error is now fixed and RRH (between) remains greater than 1.